[REF] apps_product_creator: Allow to reuse tests setup
